Enrollment Changes

+++ CscGlobalCaProxy/Client/Models/Price.cs
@@ -6,6 +6,6 @@ namespace Keyfactor.AnyGateway.CscGlobal.Client.Models
     public class Price : IPrice
     {
         [JsonProperty("currency")] public string Currency { get; set; }
-        [JsonProperty("total")] public int Total { get; set; }
+        [JsonProperty("total")] public decimal Total { get; set; }
     }
 }
